+++ riscv_config/checker.py
@@ -24,11 +24,20 @@ def nosset():
         return {'implemented': False}
 
 
-def nouset():
-    '''Function to check and set defaults for all fields which are dependent on
-        the presence of 'U' extension.'''
+def uset():
+    '''Function to set defaults based on presence of 'U' extension.'''
     global inp_yaml
     if 'U' in inp_yaml['ISA']:
+        return {'implemented': True}
+    else:
+        return {'implemented': False}
+
+
+def nuset():
+    '''Function to check and set defaults for all fields which are dependent on
+        the presence of 'U' extension and 'N' extension.'''
+    global inp_yaml
+    if 'U' in inp_yaml['ISA'] and 'N' in inp_yaml['ISA']:
         return {'implemented': True}
     else:
         return {'implemented': False}
@@ -54,40 +63,40 @@ def delegset():
           not ('N' in inp_yaml['ISA'] or 'S' in inp_yaml['ISA'])):
         var = False
 
-    temp = {'rv32': {'implemented': False}, 'rv64': {'implemented': False}}
+    temp = {'rv32': {'accessible': False}, 'rv64': {'accessible': False}}
     if 32 in inp_yaml['supported_xlen']:
-        temp['rv32']['implemented'] = True and var
+        temp['rv32']['accessible'] = True and var
     if 64 in inp_yaml['supported_xlen']:
-        temp['rv64']['implemented'] = True and var
+        temp['rv64']['accessible'] = True and var
     return temp
 
 
 def countset():
     global inp_yaml
-    temp = {'rv32': {'implemented': False}, 'rv64': {'implemented': False}}
+    temp = {'rv32': {'accessible': False}, 'rv64': {'accessible': False}}
     if 'S' in inp_yaml['ISA'] or 'U' in inp_yaml["ISA"]:
         if 32 in inp_yaml['supported_xlen']:
-            temp['rv32']['implemented'] = True
+            temp['rv32']['accessible'] = True
         if 64 in inp_yaml['supported_xlen']:
-            temp['rv64']['implemented'] = True
+            temp['rv64']['accessible'] = True
     return temp
 
 
 def regset():
     global inp_yaml
-    temp = {'rv32': {'implemented': False}, 'rv64': {'implemented': False}}
+    temp = {'rv32': {'accessible': False}, 'rv64': {'accessible': False}}
     if 32 in inp_yaml['supported_xlen']:
-        temp['rv32']['implemented'] = True
+        temp['rv32']['accessible'] = True
     if 64 in inp_yaml['supported_xlen']:
-        temp['rv64']['implemented'] = True
+        temp['rv64']['accessible'] = True
     return temp
 
 
 def counterhset():
     global inp_yaml
-    temp = {'rv32': {'implemented': False}, 'rv64': {'implemented': False}}
+    temp = {'rv32': {'accessible': False}, 'rv64': {'accessible': False}}
     if 32 in inp_yaml['supported_xlen']:
-        temp['rv32']['implemented'] = True
+        temp['rv32']['accessible'] = True
     return temp
 
 
@@ -204,45 +213,48 @@ def add_def_setters(schema_yaml):
     schema_yaml['mcounteren']['default_setter'] = lambda doc: countset()
 
     schema_yaml['mcause']['default_setter'] = regsetter
-    usetter = lambda doc: nouset()
+    nusetter = lambda doc: nuset()
     schema_yaml['mstatus']['schema']['rv32']['schema']['uie'][
-        'default_setter'] = usetter
+        'default_setter'] = nusetter
     schema_yaml['mstatus']['schema']['rv64']['schema']['uie'][
-        'default_setter'] = usetter
+        'default_setter'] = nusetter
     schema_yaml['mstatus']['schema']['rv32']['schema']['upie'][
-        'default_setter'] = usetter
+        'default_setter'] = nusetter
     schema_yaml['mstatus']['schema']['rv64']['schema']['upie'][
-        'default_setter'] = usetter
+        'default_setter'] = nusetter
+
+    usetter = lambda doc: uset()
     schema_yaml['mstatus']['schema']['rv32']['schema']['mprv'][
         'default_setter'] = usetter
     schema_yaml['mstatus']['schema']['rv64']['schema']['mprv'][
         'default_setter'] = usetter
     schema_yaml['mstatus']['schema']['rv64']['schema']['uxl'][
         'default_setter'] = usetter
+
     schema_yaml['mip']['schema']['rv32']['schema']['ueip'][
-        'default_setter'] = usetter
+        'default_setter'] = nusetter
     schema_yaml['mip']['schema']['rv64']['schema']['ueip'][
-        'default_setter'] = usetter
+        'default_setter'] = nusetter
     schema_yaml['mip']['schema']['rv32']['schema']['utip'][
-        'default_setter'] = usetter
+        'default_setter'] = nusetter
     schema_yaml['mip']['schema']['rv64']['schema']['utip'][
-        'default_setter'] = usetter
+        'default_setter'] = nusetter
     schema_yaml['mip']['schema']['rv32']['schema']['usip'][
-        'default_setter'] = usetter
+        'default_setter'] = nusetter
     schema_yaml['mip']['schema']['rv64']['schema']['usip'][
-        'default_setter'] = usetter
+        'default_setter'] = nusetter
     schema_yaml['mie']['schema']['rv32']['schema']['ueie'][
-        'default_setter'] = usetter
+        'default_setter'] = nusetter
     schema_yaml['mie']['schema']['rv64']['schema']['ueie'][
-        'default_setter'] = usetter
+        'default_setter'] = nusetter
     schema_yaml['mie']['schema']['rv32']['schema']['utie'][
-        'default_setter'] = usetter
+        'default_setter'] = nusetter
     schema_yaml['mie']['schema']['rv64']['schema']['utie'][
-        'default_setter'] = usetter
+        'default_setter'] = nusetter
     schema_yaml['mie']['schema']['rv32']['schema']['usie'][
-        'default_setter'] = usetter
+        'default_setter'] = nusetter
     schema_yaml['mie']['schema']['rv64']['schema']['usie'][
-        'default_setter'] = usetter
+        'default_setter'] = nusetter
 
     ssetter = lambda doc: nosset()
     schema_yaml['mstatus']['schema']['rv32']['schema']['sie'][
@@ -299,7 +311,6 @@ def add_def_setters(schema_yaml):
         'default_setter'] = ssetter
     schema_yaml['mie']['schema']['rv64']['schema']['ssie'][
         'default_setter'] = ssetter
-
     twsetter = lambda doc: twset()
     schema_yaml['mstatus']['schema']['rv32']['schema']['tw'][
         'default_setter'] = twsetter
@@ -323,11 +334,11 @@ def trim(foo):
     '''
     keys = foo.keys()
     if 'rv32' in keys:
-        if not foo['rv32']['implemented']:
-            foo['rv32'] = {'implemented': False}
+        if not foo['rv32']['accessible']:
+            foo['rv32'] = {'accessible': False}
     if 'rv64' in keys:
-        if not foo['rv64']['implemented']:
-            foo['rv64'] = {'implemented': False}
+        if not foo['rv64']['accessible']:
+            foo['rv64'] = {'accessible': False}
     if 'implemented' in keys:
         if not foo['implemented']:
             temp = foo
@@ -362,7 +373,7 @@ def groupc(test_list):
 def get_fields(node, bitwidth):
     fields = list(
         set(node.keys()) -
-        set(['fields', 'msb', 'lsb', 'implemented', 'shadow', 'type']))
+        set(['fields', 'msb', 'lsb', 'accessible', 'shadow', 'type']))
     if not fields:
         return fields
     bits = set(range(bitwidth))
@@ -380,18 +391,18 @@ def check_reset_fill_fields(spec):
     errors = {}
     for node in spec:
         if isinstance(spec[node], dict):
-            if spec[node]['rv32']['implemented']:
+            if spec[node]['rv32']['accessible']:
                 spec[node]['rv32']['fields'] = get_fields(
                     spec[node]['rv32'], 32)
-            if spec[node]['rv64']['implemented']:
+            if spec[node]['rv64']['accessible']:
                 spec[node]['rv64']['fields'] = get_fields(
                     spec[node]['rv64'], 64)
             if 'reset-val' in spec[node].keys():
                 reset_val = spec[node]['reset-val']
-                if spec[node]['rv64']['implemented']:
+                if spec[node]['rv64']['accessible']:
                     field_desc = spec[node]['rv64']
                     bit_len = 64
-                elif spec[node]['rv32']['implemented']:
+                elif spec[node]['rv32']['accessible']:
                     field_desc = spec[node]['rv32']
                     bit_len = 32
                 else:
@@ -401,7 +412,28 @@ def check_reset_fill_fields(spec):
                     desc = field_desc['type']
                     keys = desc.keys()
                     if 'wlrl' in keys:
-                        if reset_val not in desc['wlrl']:
+                        res = False
+                        for entry in desc['wlrl']:
+                            if ":" in entry:
+                                low, high = entry.split(":")
+                                if "x" in low:
+                                    low = int(low, base=16)
+                                    high = int(high, base=16)
+                                else:
+                                    low = int(low)
+                                    high = int(high)
+                                if reset_val >= low and reset_val <= high:
+                                    res = True
+                                    break
+                            else:
+                                if "x" in entry:
+                                    val = int(entry, base=16)
+                                else:
+                                    val = int(entry)
+                                if val == reset_val:
+                                    res = True
+                                    break
+                        if not res:
                             error.append(
                                 "Reset value doesnt match the 'wlrl' description for the register."
                             )
@@ -438,51 +470,102 @@ def check_reset_fill_fields(spec):
                     bin_str = bin(reset_val)[2:].zfill(bit_len)
                     for field in field_desc['fields']:
                         if isinstance(field, list):
-                            continue
-                        test_val = int(
-                            bin_str[bit_len - 1 -
-                                    field_desc[field]['msb']:bit_len -
-                                    field_desc[field]['lsb']],
-                            base=2)
-                        desc = field_desc[field]['type']
-                        keys = desc.keys()
-                        if 'wlrl' in keys:
-                            if test_val not in desc['wlrl']:
-                                self._error(
-                                    field, "Reset value for " + field +
-                                    " doesnt match the 'wlrl' description.")
-                        elif 'ro_constant' in keys:
-                            if test_val not in desc['ro_constant']:
-                                error.append(
-                                    "Reset value for " + field +
-                                    " doesnt match the 'ro_constant' description."
-                                )
-                        elif 'ro_variable' in keys:
-                            pass
-                        elif "warl" in keys:
-                            warl = (warl_interpreter(desc['warl']))
-                            deps = warl.dependencies()
-                            dep_vals = []
-                            for dep in deps:
-                                reg = dep.split("::")
-                                if len(reg) == 1:
-                                    dep_vals.append(spec[reg[0]]['reset-val'])
-                                else:
-                                    bin_str = bin(spec[
-                                        reg[0]]['reset-val'])[2:].zfill(bit_len)
-                                    dep_vals.append(
-                                        int(bin_str[
-                                            bit_len - 1 -
-                                            spec[reg[0]]['rv{}'.format(bit_len)]
-                                            [reg[1]]['msb']:bit_len -
-                                            spec[reg[0]]['rv{}'.format(
-                                                bit_len)][reg[1]]['lsb']],
-                                            base=2))
-                            if (warl.islegal(hex(test_val)[2:], dep_vals) !=
-                                    True):
-                                error.append(
-                                    "Reset value for " + field +
-                                    " doesnt match the 'warl' description.")
+                            for entry in field:
+                                if len(entry) == 2:
+                                    test_val = int(
+                                        bin_str[bit_len - 1 - entry[1]:bit_len -
+                                                entry[0]],
+                                        base=2)
+                                    if test_val != 0:
+                                        error.append("WPRI bits " +
+                                                     str(entry[0]) + " to " +
+                                                     str(entry[1]) +
+                                                     " should be zero.")
+                                elif len(entry) == 1:
+                                    test_val = int(bin_str[bit_len - 1 -
+                                                           entry[0]],
+                                                   base=2)
+                                    if test_val != 0:
+                                        error.append("WPRI bit " +
+                                                     str(entry[0]) +
+                                                     " should be zero.")
+                        else:
+                            test_val = int(
+                                bin_str[bit_len - 1 -
+                                        field_desc[field]['msb']:bit_len -
+                                        field_desc[field]['lsb']],
+                                base=2)
+                            if field_desc[field]['implemented']:
+                                desc = field_desc[field]['type']
+                                keys = desc.keys()
+                                if 'wlrl' in keys:
+                                    res = False
+                                    for entry in desc['wlrl']:
+                                        if ":" in entry:
+                                            low, high = entry.split(":")
+                                            if "x" in low:
+                                                low = int(low, base=16)
+                                                high = int(high, base=16)
+                                            else:
+                                                low = int(low)
+                                                high = int(high)
+                                            if test_val >= low and test_val <= high:
+                                                res = True
+                                                break
+                                        else:
+                                            if "x" in entry:
+                                                val = int(entry, base=16)
+                                            else:
+                                                val = int(entry)
+                                            if val == test_val:
+                                                res = True
+                                                break
+                                    if not res:
+                                        error.append(
+                                            "Reset value for " + field +
+                                            " doesnt match the 'wlrl' description."
+                                        )
+                                elif 'ro_constant' in keys:
+                                    if test_val not in desc['ro_constant']:
+                                        error.append(
+                                            "Reset value for " + field +
+                                            " doesnt match the 'ro_constant' description."
+                                        )
+                                elif 'ro_variable' in keys:
+                                    pass
+                                elif "warl" in keys:
+                                    warl = (warl_interpreter(desc['warl']))
+                                    deps = warl.dependencies()
+                                    dep_vals = []
+                                    for dep in deps:
+                                        reg = dep.split("::")
+                                        if len(reg) == 1:
+                                            dep_vals.append(
+                                                spec[reg[0]]['reset-val'])
+                                        else:
+                                            bin_str = bin(
+                                                spec[reg[0]]['reset-val']
+                                            )[2:].zfill(bit_len)
+                                            dep_vals.append(
+                                                int(bin_str[
+                                                    bit_len - 1 -
+                                                    spec[reg[0]]['rv{}'.format(
+                                                        bit_len
+                                                    )][reg[1]]['msb']:bit_len -
+                                                    spec[reg[0]]['rv{}'.format(
+                                                        bit_len
+                                                    )][reg[1]]['lsb']],
+                                                    base=2))
+                                    if (warl.islegal(
+                                            hex(test_val)[2:], dep_vals) !=
+                                            True):
+                                        error.append(
+                                            "Reset value for " + field +
+                                            " doesnt match the 'warl' description."
+                                        )
+                            elif test_val != 0:
+                                error.append("Reset value for unimplemented " +
+                                             field + " cannot be non zero.")
                 if error:
                     errors[node] = error
     return spec, errors